chore: convert useSettingsMenu to TS
