--- packages/core/admin/admin/src/hooks/useSettingsMenu/utils/formatLinks.js
@@ -1,12 +1,0 @@
-const formatLinks = (menu) => {
-  return menu.map((menuSection) => {
-    const formattedLinks = menuSection.links.map((link) => ({
-      ...link,
-      isDisplayed: false,
-    }));
-
-    return { ...menuSection, links: formattedLinks };
-  });
-};
-
-export default formatLinks;
