--- packages/core/admin/admin/src/hooks/useSettingsMenu/utils/sortLinks.js
@@ -1,5 +1,0 @@
-import sortBy from 'lodash/sortBy';
-
-const sortLinks = (links) => sortBy(links, (link) => link.id);
-
-export default sortLinks;
