+++ packages/core/admin/admin/src/pages/SettingsPage/tests/index.test.js
@@ -10,10 +10,13 @@ import { Route, Router } from 'react-router-dom';
 import { SettingsPage } from '..';
 import { Theme } from '../../../components/Theme';
 import { ThemeToggleProvider } from '../../../components/ThemeToggleProvider';
-import { useSettingsMenu } from '../../../hooks';
+import { useSettingsMenu } from '../../../hooks/useSettingsMenu';
+
+jest.mock('../../../hooks/useSettingsMenu', () => ({
+  useSettingsMenu: jest.fn(() => ({ isLoading: false, menu: [] })),
+}));
 
 jest.mock('../../../hooks', () => ({
-  useSettingsMenu: jest.fn(() => ({ isLoading: false, menu: [] })),
   useAppInfo: jest.fn(() => ({ shouldUpdateStrapi: false })),
   useThemeToggle: jest.fn(() => ({ currentTheme: 'light', themes: { light: lightTheme } })),
 }));
