--- packages/core/admin/ee/admin/constants.js
@@ -1,19 +1,0 @@
-export const ADMIN_PERMISSIONS_EE = {
-  settings: {
-    auditLogs: {
-      main: [{ action: 'admin::audit-logs.read', subject: null }],
-      read: [{ action: 'admin::audit-logs.read', subject: null }],
-    },
-    'review-workflows': {
-      main: [{ action: 'admin::review-workflows.read', subject: null }],
-      create: [{ action: 'admin::review-workflows.create', subject: null }],
-      delete: [{ action: 'admin::review-workflows.delete', subject: null }],
-      update: [{ action: 'admin::review-workflows.update', subject: null }],
-    },
-    sso: {
-      main: [{ action: 'admin::provider-login.read', subject: null }],
-      read: [{ action: 'admin::provider-login.read', subject: null }],
-      update: [{ action: 'admin::provider-login.update', subject: null }],
-    },
-  },
-};
