+++ packages/core/admin/ee/admin/constants.ts
@@ -1,0 +1,58 @@
+export const ADMIN_PERMISSIONS_EE = {
+  settings: {
+    auditLogs: {
+      main: [{ action: 'admin::audit-logs.read', subject: null }],
+      read: [{ action: 'admin::audit-logs.read', subject: null }],
+    },
+    'review-workflows': {
+      main: [{ action: 'admin::review-workflows.read', subject: null }],
+      create: [{ action: 'admin::review-workflows.create', subject: null }],
+      delete: [{ action: 'admin::review-workflows.delete', subject: null }],
+      update: [{ action: 'admin::review-workflows.update', subject: null }],
+    },
+    sso: {
+      main: [{ action: 'admin::provider-login.read', subject: null }],
+      read: [{ action: 'admin::provider-login.read', subject: null }],
+      update: [{ action: 'admin::provider-login.update', subject: null }],
+    },
+  },
+};
+
+export const SETTINGS_MENU_LINKS_EE = {
+  global: [
+    ...(window.strapi.features.isEnabled(window.strapi.features.SSO)
+      ? [
+          {
+            intlLabel: { id: 'Settings.sso.title', defaultMessage: 'Single Sign-On' },
+            to: '/settings/single-sign-on',
+            id: 'sso',
+          },
+        ]
+      : []),
+
+    ...(window.strapi.features.isEnabled(window.strapi.features.REVIEW_WORKFLOWS)
+      ? [
+          {
+            intlLabel: {
+              id: 'Settings.review-workflows.page.title',
+              defaultMessage: 'Review Workflows',
+            },
+            to: '/settings/review-workflows',
+            id: 'review-workflows',
+          },
+        ]
+      : []),
+  ],
+
+  admin: [
+    ...(window.strapi.features.isEnabled(window.strapi.features.AUDIT_LOGS)
+      ? [
+          {
+            intlLabel: { id: 'global.auditLogs', defaultMessage: 'Audit Logs' },
+            to: '/settings/audit-logs?pageSize=50&page=1&sort=date:DESC',
+            id: 'auditLogs',
+          },
+        ]
+      : []),
+  ],
+};
